// frontend/lib/api/generated/model/informeConformidadPagoDetailOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */
import type { InformeConformidadDetalleOut } from './informeConformidadDetalleOut';

export interface InformeConformidadPagoDetailOut {
  informe_id: number;
  numero_informe: string;
  orden_compra_id: number;
  monto_conforme_total: number;
  monto_retenido_total: number;
  monto_penalidad_total: number;
  generado_por_id: number;
  estado: string;
  fecha_envio: string;
  fecha_pago: string | null;
  detalle: InformeConformidadDetalleOut[];
}