// frontend/lib/api/generated/model/ordenCompraOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface OrdenCompraOut {
  orden_compra_id: number;
  numero_oc: string;
  contrato_id: number;
  periodo_mes: string;
  estado: string;
  responsable_id: number;
  creado_en: string;
}